addadminsay

+++ index.js
@@ -49,7 +49,6 @@ bot.on('ready', () => {
 });
 
 
-
 bot.on('messageCreate', (message) => {
   if (message.author.bot) return; // Ne répondez pas aux messages des bots
   if (!message.content.startsWith(config.prefix)) return; // Vérifiez s'il commence par le préfixe
@@ -58,8 +57,18 @@ bot.on('messageCreate', (message) => {
   const command = args.shift().toLowerCase();
 
   if (command === 'say') {
+    // Vérifiez si l'utilisateur est un administrateur
+    if (!message.member.permissions.has('ADMINISTRATOR')) {
+      return message.reply('Seuls les administrateurs sont autorisés à utiliser cette commande.');
+    } 
+
     // Supprime la commande de l'utilisateur
     message.delete();
+
+    // Vérifiez si l'utilisateur est également un administrateur pour utiliser args.join
+    if (!message.member.permissions.has('ADMINISTRATOR')) {
+      return message.reply('Seuls les administrateurs sont autorisés à utiliser cette partie de la commande.');
+    }
 
     // Récupère le message de l'utilisateur, en excluant le préfixe
     const userMessage = args.join(' ');
@@ -71,6 +80,9 @@ bot.on('messageCreate', (message) => {
     message.channel.send(userMessage);
   }
 });
+
+
+
